requirements.txt
Django<5.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
django-jazzmin==2.6.0
django-jsonform>=2.22.0
python-dotenv>=1.0.0
# pandas>=2.0.0  <-- 削除（または行頭に#をつけてコメントアウト）